requirements.txt
inquirer>=3.1.3
google_adk>=0.4.0
litellm>=1.68.0
python-dotenv>=1.0.0
boto3>=1.38.9
pyodbc>=4.0.35
questionary>=2.0.1
sqlparse>=0.4.4
